NIFI-9806 Introduce ConfigurableExtensionDefinition and VersionedConfigurableExtension
