+++ c2/c2-protocol/c2-protocol-component-api/src/main/java/org/apache/nifi/c2/protocol/component/api/ControllerServiceDefinition.java
@@ -18,39 +18,9 @@
 package org.apache.nifi.c2.protocol.component.api;
 
 import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.Map;
 
 @ApiModel
-public class ControllerServiceDefinition extends ExtensionComponent implements ConfigurableComponentDefinition {
+public class ControllerServiceDefinition extends ConfigurableExtensionDefinition {
     private static final long serialVersionUID = 1L;
-
-    private Map<String, PropertyDescriptor> propertyDescriptors;
-    private boolean supportsDynamicProperties;
-
-    @Override
-    @ApiModelProperty("Descriptions of configuration properties applicable to this controller service")
-    public Map<String, PropertyDescriptor> getPropertyDescriptors() {
-        return (propertyDescriptors != null ? Collections.unmodifiableMap(propertyDescriptors) : null);
-    }
-
-    @Override
-    public void setPropertyDescriptors(LinkedHashMap<String, PropertyDescriptor> propertyDescriptors) {
-        this.propertyDescriptors = propertyDescriptors;
-    }
-
-    @Override
-    @ApiModelProperty("Whether or not this processor makes use of dynamic (user-set) properties")
-    public boolean getSupportsDynamicProperties() {
-        return supportsDynamicProperties;
-    }
-
-    @Override
-    public void setSupportsDynamicProperties(boolean supportsDynamicProperties) {
-        this.supportsDynamicProperties = supportsDynamicProperties;
-    }
 
 }
