+++ c2/c2-protocol/c2-protocol-component-api/src/main/java/org/apache/nifi/c2/protocol/component/api/ProcessorDefinition.java
@@ -22,18 +22,14 @@ import io.swagger.annotations.ApiModelProperty;
 import org.apache.nifi.annotation.behavior.InputRequirement;
 
 import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
 @ApiModel
-public class ProcessorDefinition extends ExtensionComponent implements ConfigurableComponentDefinition {
+public class ProcessorDefinition extends ConfigurableExtensionDefinition {
     private static final long serialVersionUID = 1L;
 
-    private Map<String, PropertyDescriptor> propertyDescriptors;
-    private boolean supportsDynamicProperties;
     private InputRequirement.Requirement inputRequirement;
-
     private List<Relationship> supportedRelationships;
     private boolean supportsDynamicRelationships;
 
@@ -53,28 +49,6 @@ public class ProcessorDefinition extends ExtensionComponent implements Configura
     private String defaultPenaltyDuration;
     private String defaultYieldDuration;
     private String defaultBulletinLevel;
-
-    @Override
-    @ApiModelProperty("Descriptions of configuration properties applicable to this processor.")
-    public Map<String, PropertyDescriptor> getPropertyDescriptors() {
-        return (propertyDescriptors != null ? Collections.unmodifiableMap(propertyDescriptors) : null);
-    }
-
-    @Override
-    public void setPropertyDescriptors(LinkedHashMap<String, PropertyDescriptor> propertyDescriptors) {
-        this.propertyDescriptors = propertyDescriptors;
-    }
-
-    @Override
-    @ApiModelProperty("Whether or not this processor makes use of dynamic (user-set) properties.")
-    public boolean getSupportsDynamicProperties() {
-        return supportsDynamicProperties;
-    }
-
-    @Override
-    public void setSupportsDynamicProperties(boolean supportsDynamicProperties) {
-        this.supportsDynamicProperties = supportsDynamicProperties;
-    }
 
     @ApiModelProperty("Any input requirements this processor has.")
     public InputRequirement.Requirement getInputRequirement() {
