+++ nifi-api/src/main/java/org/apache/nifi/flow/VersionedControllerService.java
@@ -20,41 +20,13 @@ package org.apache.nifi.flow;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.util.List;
-import java.util.Map;
 
-public class VersionedControllerService extends VersionedComponent
-        implements VersionedConfigurableComponent, VersionedExtensionComponent {
+public class VersionedControllerService extends VersionedConfigurableExtension {
 
-    private String type;
-    private Bundle bundle;
     private List<ControllerServiceAPI> controllerServiceApis;
 
-    private Map<String, String> properties;
-    private Map<String, VersionedPropertyDescriptor> propertyDescriptors;
     private String annotationData;
     private ScheduledState scheduledState;
-
-    @Override
-    @ApiModelProperty(value = "The type of the controller service.")
-    public String getType() {
-        return type;
-    }
-
-    @Override
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    @Override
-    @ApiModelProperty(value = "The details of the artifact that bundled this controller service type.")
-    public Bundle getBundle() {
-        return bundle;
-    }
-
-    @Override
-    public void setBundle(Bundle bundle) {
-        this.bundle = bundle;
-    }
 
     @ApiModelProperty(value = "Lists the APIs this Controller Service implements.")
     public List<ControllerServiceAPI> getControllerServiceApis() {
@@ -63,28 +35,6 @@ public class VersionedControllerService extends VersionedComponent
 
     public void setControllerServiceApis(List<ControllerServiceAPI> controllerServiceApis) {
         this.controllerServiceApis = controllerServiceApis;
-    }
-
-    @Override
-    @ApiModelProperty(value = "The properties of the controller service.")
-    public Map<String, String> getProperties() {
-        return properties;
-    }
-
-    @Override
-    public void setProperties(Map<String, String> properties) {
-        this.properties = properties;
-    }
-
-    @Override
-    @ApiModelProperty("The property descriptors for the controller service.")
-    public Map<String, VersionedPropertyDescriptor> getPropertyDescriptors() {
-        return propertyDescriptors;
-    }
-
-    @Override
-    public void setPropertyDescriptors(Map<String, VersionedPropertyDescriptor> propertyDescriptors) {
-        this.propertyDescriptors = propertyDescriptors;
     }
 
     @ApiModelProperty(value = "The annotation for the controller service. This is how the custom UI relays configuration to the controller service.")
